refactor: singleton for node service
