+++ packages/neuron-wallet/src/services/sync-blocks.ts
@@ -6,7 +6,7 @@ import { Script, OutPoint, Cell, Input, Transaction, Block, BlockHeader } from '
 import TransactionsService from './transactions'
 import OutputEntity from '../entities/output'
 import SyncInfoEntity from '../entities/sync-info'
-import nodeService from '../startup/node-service'
+import { nodeService } from './node'
 import LockUtils from '../utils/lock-utils'
 import TypeConvert from '../app-types/type-convert'
 import { NetworkWithID } from './networks'
