+++ packages/neuron-wallet/src/services/wallets.ts
@@ -5,7 +5,7 @@ import TransactionsService from './transactions'
 import Key, { Addresses } from '../keys/key'
 import { Keystore } from '../keys/keystore'
 import Store from '../utils/store'
-import nodeService from '../startup/node-service'
+import { nodeService } from './node'
 import fileService from '../startup/file-service'
 import LockUtils from '../utils/lock-utils'
 import env from '../env'
