+++ packages/neuron-wallet/src/startup/init-app.ts
@@ -2,7 +2,7 @@ import { distinctUntilChanged } from 'rxjs/operators'
 import controllers from '../controllers'
 import windowManage from '../utils/window-manage'
 import Router from '../router'
-import nodeService from './node-service'
+import { nodeService } from '../services/node'
 import { Channel } from '../utils/const'
 import logger from '../utils/logger'
 import app from '../app'
