--- packages/neuron-wallet/src/startup/node-service.ts
@@ -1,3 +1,0 @@
-import NodeService from '../services/node'
-
-export default new NodeService()
